separate old and vlen implementation
