+++ tests/test_parser_vlen.py
@@ -1,11 +1,11 @@
 import numpy as np
-from dstparser import parse_dst_file, parse_dst_file_tax4
+from dstparser import parse_dst_file_vlen
 from time import time
 from dstparser.cli.cli import parse_config
 import sys
 
 
-def test_parser(dst_file, print_read_data=False, use_ta_x4=False, add_xmax=False):
+def test_parser(dst_file, print_read_data=False, add_xmax=False):
 
     if len(sys.argv) > 1:
         config = parse_config(sys.argv[1])
@@ -18,30 +18,23 @@ def test_parser(dst_file, print_read_data=False, use_ta_x4=False, add_xmax=False
     # paths to the directories with data is dstparser.paths module
     # data = parse_dst_file(dst_file, up_low_traces=True)
 
-    if use_ta_x4:
-        convert_func = parse_dst_file_tax4
-    else:
-        convert_func = parse_dst_file
-
+    xmax_reader = None
     if add_xmax:
         from dstparser.xmax_reader import XmaxReader
 
         xmax_dir = "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/qgsii04proton/080417_160603/Em1_bsdinfo"
         xmax_reader = XmaxReader(xmax_dir, "**/DAT*_xmax.txt", "QGSJetII-04")
-    else:
-        xmax_reader = None
 
-    data = convert_func(
+    data = parse_dst_file_vlen(
         dst_file,
-        ntile=7,
         xmax_reader=xmax_reader,
-        avg_traces=False,
         add_shower_params=True,
         add_standard_recon=True,
         config=config,
     )
 
     end = time()
+    # print(data)
 
     if print_read_data:
         print(f"\nConverted arrays:\n---")
@@ -82,7 +75,4 @@ if __name__ == "__main__":
     # dst_file = "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/tax4/qgsii04proton/north/221101to240124/DAT010019_gea.rufldf.dst.gz"
 
     # dst_file = "/ceph/work/SATORI/projects/TA-ASIoP/tasdmc_dstbank/qgsii04nitrogen/080417_160603/Em1_bsdinfo/DAT081325_gea.rufldf.dst.gz"
-    # !If you want to use TAx4 format, set use_ta_x4=True
-    # !If you want to use TA format, set use_ta_x4=False
-    # !By default, use_ta_x4=False
-    test_parser(dst_file, print_read_data=True, use_ta_x4=False, add_xmax=False)
+    test_parser(dst_file, print_read_data=True, add_xmax=True)
